access auth logics

+++ backend/todo/todo_view_elements/group.py
@@ -1,4 +1,5 @@
 from ..models import Group
+from ..responses import *
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -7,6 +8,9 @@ from rest_framework.response import Response
 class GroupAPI(APIView):
 
     def post(self, request):
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
+        
         data = request.data
 
         Group.objects.create(
@@ -16,8 +20,10 @@ class GroupAPI(APIView):
         return Response(data)
 
     def patch(self, request):
-        data = request.data
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
 
+        data = request.data
         Group.objects.filter(id=data['id']).update(
             name=data['name'],
             color=data['color']
@@ -25,6 +31,9 @@ class GroupAPI(APIView):
         return Response(data)
 
     def delete(self, request):
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
+        
         data = request.data
         Group.objects.delete(id=data['id'])
         return Response(data)
