+++ backend/todo/todo_view_elements/task.py
@@ -1,4 +1,5 @@
 from ..models import Task
+from ..responses import *
 
 from rest_framework.views import APIView
 from rest_framework.response import Response
@@ -7,8 +8,10 @@ from rest_framework.response import Response
 class TaskAPI(APIView):
 
     def post(self, request):
-        data = request.data
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
 
+        data = request.data
         Task.objects.create(
             name=data['name'],
             group_id=data['group_id'],
@@ -18,8 +21,10 @@ class TaskAPI(APIView):
         return Response(data)
 
     def patch(self, request):
-        data = request.data
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
 
+        data = request.data
         Task.objects.filter(id=data['id']).update(
             name=data['name'],
             color=data['color'],
@@ -29,6 +34,9 @@ class TaskAPI(APIView):
         return Response(data)
 
     def delete(self, request):
+        if request.user.is_anonymous:
+            return UNAUTHORIZED_401
+        
         data = request.data
         Task.objects.delete(id=data['id'])
         return Response(data)
